requirements.txt
Django==4.0.5
djangorestframework==3.13.1
requests==2.27.1
djangorestframework-simplejwt==5.2.0